requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
forex-python==1.6
idna==2.10
lxml==4.6.5
multitasking==0.0.9
numpy==1.21.0
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0.1
requests==2.25.1
scipy==1.7.0rc1
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
yfinance==0.1.60
